reorganized source code
