+++ src/eTracker/command/Command.java
@@ -1,0 +1,17 @@
+package eTracker.command;
+
+public final class Command {
+	
+		// command in this Expense tracker program
+		public final static String ADD = "add";
+		public final static String UPDATE = "update";
+		public final static String DELETE = "delete";
+		public final static String LIST = "list";
+		public final static String SUMMARY = "summary";
+		public final static String SETBUDGET = "set-budget";
+		public final static String LISTBUDGET = "list-budget";
+		public final static String GETBUDGET = "get-budget";
+
+		public final static String EXIT = "exit";
+		public final static String SAVE = "save";
+}
