+++ src/eTracker/command/behavior/Exit.java
@@ -1,0 +1,5 @@
+package eTracker.command.behavior;
+
+public class Exit {
+
+}
