+++ src/eTracker/object/Expense.java
@@ -1,4 +1,4 @@
-package expense_tracker.object;
+package eTracker.object;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
